% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DockerRunORCA.R
\name{docker.application.run}
\alias{docker.application.run}
\title{Run docker container}
\usage{
docker.application.run(
  port = 3838,
  image = "qbioturin/orca-shiny-application:latest",
  debug = FALSE
)
}
\description{
This function runs the Shiny app using its Docker image. It pulls the image if not available,
opens the browser, and monitors for errors.
}
\examples{
\dontrun{
  docker.application.run()
}
}
\author{
Pernice Simone
}
