% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispaly_data.R
\name{InteGreat.run}
\alias{InteGreat.run}
\title{Run InteGreat}
\usage{
InteGreat.run()
}
\description{
function to lunch the InteGreat shiny application.
}
\examples{
\dontrun{
InteGreat.run()
}
}
\author{
Pernice Simone Tortarolo Dora
}
