\name{karate}
\docType{data}
\alias{karate}
\title{Karate Club network}
\description{
  The network representing the friendships between members of a
  university-based karate club, which was originally used to
  model the fission process with a mathematical model. The network
  consists of 34 nodes (club members) and 77 edges (friendships).
}
\usage{karate}
\format{A data frame with 77 observations and 2 columns.}
\source{https://networkdata.ics.uci.edu/data.php?id=105}
\references{
  W. W. Zachary (1977)
  An information flow model for conflict and fission in small groups, 
  \emph{Journal of Anthropological Research} 33(4), 452-473.
}
\keyword{datasets}