% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runORCA.R
\name{ORCA.run}
\alias{ORCA.run}
\title{Run ORCA}
\usage{
ORCA.run(
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\description{
function to lunch the ORCA shiny application.
}
\examples{
\dontrun{
ORCA.run()
}
}
\author{
Beccuti Marco, Pernice Simone, Tortarolo Dora
}
