\encoding{utf8}
\name{petersen}
\docType{data}
\alias{petersen}
\title{Petersen graph}
\description{
  The Petersen graph is a graph with 10 vertices and 15 edges, usually
  drawn as a pentagram within a pentagon.
}
\usage{petersen}
\format{A data frame with 15 observations and 2 columns.}
% \source{https://networkdata.ics.uci.edu/data.php?id=105}
\references{
  J. Petersen
  Sur la théorème de Tait.
  \emph{L'Intermédiare des Math}. 5, 225-227.}
\keyword{datasets}