\name{yeast}
\docType{data}
\alias{yeast}
\title{Karate Club network}
\description{
  The network representing interactions between 2361 proteins.
}
\usage{karate}
\format{A data frame with 6646 observations and 2 columns.}
\source{http://vlado.fmf.uni-lj.si/pub/networks/data/bio/Yeast/Yeast.htm}
\references{
	S. Sun, L. Ling, N. Zhang, G. Li and R. Chen (2003)
	Topological structure analysis of the protein-protein interaction
	network in budding yeast,
	\emph{Nucleic Acids Research}, 31(9), 2443-2450.
}
\keyword{datasets}