% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DockerRunApplication.R
\name{docker.application.run}
\alias{docker.application.run}
\title{Run docker container}
\usage{
docker.application.run()
}
\description{
This is the function to run the application through its docker image
}
\examples{
\dontrun{
    #running runDocker
     docker.run(params=NULL, changeUID=TRUE, debug=FALSE)

}
}
\author{
Pernice Simone
}
