% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispaly_data.R
\name{ORCA.run}
\alias{ORCA.run}
\title{Run ORCA}
\usage{
ORCA.run()
}
\description{
function to lunch the ORCA shiny application.
}
\examples{
\dontrun{
ORCA.run()
}
}
\author{
Beccuti Marco, Pernice Simone, Tortarolo Dora
}
