\name{usastates}
\docType{data}
\alias{usastates}
\title{Contiguous USA Graph}
\description{
  The nodes of Contiguous USA Graph represents the 48 contiguous states
  of the USA. Edges connect the pair connected by at least on driveable
  road.
}
\usage{usastates}
\format{A data frame with 107 observations and 2 columns.}
\source{http://www-cs-staff.stanford.edu/~uno/contiguous-usa.dat}
\references{
	D. E. Knuth (2008)
	\emph{The Art of Computer Programming, Volume 4, Fascicle 0:
	Introduction to Combinatorial Functions and Boolean Functions},
	p. 15.
	Upper Saddle River, NJ: Addison-Wesley.
}
\keyword{datasets}